Using existing constant class + simplifying implementation + updating existing method
