+++ plugins/woocommerce/includes/gateways/paypal/class-wc-gateway-paypal.php
@@ -522,7 +522,7 @@ class WC_Gateway_Paypal extends WC_Payment_Gateway {
 			get_woocommerce_currency(),
 			apply_filters(
 				'woocommerce_paypal_supported_currencies',
-				array( 'AUD', 'BRL', 'CAD', 'MXN', 'NZD', 'HKD', 'SGD', 'USD', 'EUR', 'JPY', 'TRY', 'NOK', 'CZK', 'DKK', 'HUF', 'ILS', 'MYR', 'PHP', 'PLN', 'SEK', 'CHF', 'TWD', 'THB', 'GBP', 'RMB', 'RUB', 'INR' )
+				WC_Gateway_Paypal_Constants::SUPPORTED_CURRENCIES
 			),
 			true
 		);
