+++ plugins/woocommerce/includes/gateways/paypal/includes/class-wc-gateway-paypal-request.php
@@ -8,7 +8,6 @@
 declare(strict_types=1);
 
 use Automattic\WooCommerce\Gateways\PayPal\AddressRequirements as PayPalAddressRequirements;
-use Automattic\WooCommerce\Gateways\PayPal\CurrenciesSupported as PayPalCurrenciesSupported;
 use Automattic\WooCommerce\Utilities\NumberUtil;
 use Automattic\WooCommerce\Enums\OrderStatus;
 use Automattic\Jetpack\Connection\Client as Jetpack_Connection_Client;
@@ -415,10 +414,7 @@ class WC_Gateway_Paypal_Request {
 		$shipping_preference = $this->get_paypal_shipping_preference( $order );
 
 		// Check if the order currency is supported by PayPal.
-		// phpcs:ignore Generic.Commenting.Todo.TaskFound
-		// TODO: The container call can be removed once we migrate this class to the `src` folder.
-		$currencies_supported = wc_get_container()->get( PayPalCurrenciesSupported::class )::instance();
-		if ( ! $currencies_supported->is_currency_supported( $order->get_currency() ) ) {
+		if ( ! in_array( strtoupper( $order->get_currency() ), WC_Gateway_Paypal_Constants::SUPPORTED_CURRENCIES, true ) ) {
 			throw new Exception( 'Currency is not supported by PayPal. Order ID: ' . esc_html( $order->get_id() ) );
 		}
